+++ app/shaders/d3d11_overlay_pixel.hlsl
@@ -3,8 +3,8 @@ SamplerState theSampler : register(s0);
 
 struct ShaderInput
 {
-    min16float4 pos : SV_POSITION;
-    min16float2 tex : TEXCOORD0;
+    float4 pos : SV_POSITION;
+    float2 tex : TEXCOORD0;
 };
 
 min16float4 main(ShaderInput input) : SV_TARGET
