+++ app/shaders/d3d11_vertex.hlsl
@@ -1,19 +1,19 @@
 struct ShaderInput
 {
-    min16float2 pos : POSITION;
-    min16float2 tex : TEXCOORD0;
+    float2 pos : POSITION;
+    float2 tex : TEXCOORD0;
 };
 
 struct ShaderOutput
 {
-    min16float4 pos : SV_POSITION;
-    min16float2 tex : TEXCOORD0;
+    float4 pos : SV_POSITION;
+    float2 tex : TEXCOORD0;
 };
 
 ShaderOutput main(ShaderInput input) 
 {
     ShaderOutput output;
-    output.pos = min16float4(input.pos, 0.0, 1.0);
+    output.pos = float4(input.pos, 0.0, 1.0);
     output.tex = input.tex;
     return output;
 }
